test(util): `LottoNumberGenerator` 테스트 코드 수정
작업 내용
- 테스트 클래스명을 `RandomGeneratorTest`에서 `LottoNumberGeneratorTest`로 변경
- Random API의 반환값에 대한 테스트 추가 (Random API를 사용해 무작위로 번호를 생성한다)
